# src/config/swagger.yaml
swagger: "2.0"
info:
  description: "GestIC App API documentation"
  version: "1.0.0"
  title: "GestIC API"
host: "localhost:3333"
...
paths:
...
  /offer:
    get:
      tags:
      - "Class Offer"
      summary: "Get a list of all class offers"
      produces:
      - "application/json"
      responses:
        "200":
          description: "Obtained a list of all class offers"
          schema:
            type: "array"
            produces: "application/json"
            items:
              type: "object"
              properties:
                id:
                  type: "string"
                  format: "uuid"
                name:
                  type: "string"
                code:
                  type: "string"
                codeClassroom:
                  type: "string"
                linkClassroom:
                  type: "string"
                linkMeets:
                  type: "string"
                linkWpp:
                  type: "string"
                linkTel:
                  type: "string"
        "500":
          description: "Could not obtain a list of all class offers"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    post:
      tags:
      - "Class Offer"
      summary: "Insert a class offer"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: "object"
            properties:
              name:
                type: "string"
              code:
                type: "string"
              codeClassroom:
                type: "string"
              linkClassroom:
                type: "string"
              linkMeets:
                type: "string"
              linkWpp:
                type: "string"
              linkTel:
                type: "string"
      responses:
        "200":
          description: "Inserted a class offer"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              code:
                type: "string"
              codeClassroom:
                type: "string"
              linkClassroom:
                type: "string"
              linkMeets:
                type: "string"
              linkWpp:
                type: "string"
              linkTel:
                type: "string"
        "500":
          description: "Could not insert a class offer"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    put:
      tags:
      - "Class Offer"
      summary: " Update class offer information"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          description: "Any quantity of the following parameters can be sent, but id is required."
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              code:
                type: "string"
              codeClassroom:
                type: "string"
              linkClassroom:
                type: "string"
              linkMeets:
                type: "string"
              linkWpp:
                type: "string"
              linkTel:
                type: "string"
      responses:
        "200":
          description: "Sucessfully updated a class offer"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              code:
                type: "string"
              codeClassroom:
                type: "string"
              linkClassroom:
                type: "string"
              linkMeets:
                type: "string"
              linkWpp:
                type: "string"
              linkTel:
                type: "string"
        "500":
          description: "Could not update a class offer"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    delete:
      tags:
        - "Class Offer"
      summary: "Delete all class offers"
      produces:
      - "application/json"
      responses:
        "200":
          description: "Successfully removed all class offers"
          schema:
            type: "array"
            produces: "application/json"
            items:
              type: "object"
              properties:
                id:
                  type: "string"
                  format: "uuid"
                name:
                  type: "string"
                code:
                  type: "string"
                codeClassroom:
                  type: "string"
                linkClassroom:
                  type: "string"
                linkMeets:
                  type: "string"
                linkWpp:
                  type: "string"
                linkTel:
                  type: "string"
        "500":
          description: "Could not remove all class offers"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
  /offer/:id:
    get:
      tags:
        - "Class Offer"
      summary: "Get class offer information"
      produces:
      - "application/json"
      parameters:
        - in: param
          name: id
          required: true
          type: string
          format: "uuid"
      responses:
        "200":
          description: "Obtained a class offer"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              code:
                type: "string"
              codeClassroom:
                type: "string"
              linkClassroom:
                type: "string"
              linkMeets:
                type: "string"
              linkWpp:
                type: "string"
              linkTel:
                type: "string"
        "500":
          description: "Could not obtain a class offer"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    delete:
      tags:
        - "Class Offer"
      summary: "Delete class offer"
      produces:
      - "application/json"
      parameters:
        - in: param
          name: id
          required: true
          type: string
          format: "uuid"
      responses:
        "200":
          description: "Successfully removed a class offer"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              code:
                type: "string"
              codeClassroom:
                type: "string"
              linkClassroom:
                type: "string"
              linkMeets:
                type: "string"
              linkWpp:
                type: "string"
              linkTel:
                type: "string"
        "500":
          description: "Could not remove a class offer"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
  /informative:
    get:
      tags:
      - "Informatives"
      summary: "Get a list of all informatives"
      produces:
      - "application/json"
      responses:
        "200":
          description: "Obtained a list of all informatives"
          schema:
            type: "array"
            produces: "application/json"
            items:
              type: "object"
              properties:
                id:
                  type: "string"
                  format: "uuid"
                title:
                  type: "string"
                content:
                  type: "string"
        "500":
          description: "Could not obtain a list of all informatives"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    post:
      tags:
      - "Informatives"
      summary: "Insert a informative"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: "object"
            properties:
              title:
                type: "string"
              content:
                type: "string"
      responses:
        "200":
          description: "Inserted a informative"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              title:
                type: "string"
              content:
                type: "string"
        "500":
          description: "Could not insert a informative"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    put:
      tags:
      - "Informatives"
      summary: " Update informative information"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          description: "Any quantity of the following parameters can be sent, but id is required."
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              title:
                type: "string"
              content:
                type: "string"
      responses:
        "200":
          description: "Sucessfully updated a informative"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              title:
                type: "string"
              content:
                type: "string"
        "500":
          description: "Could not update a informative"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
  /informative/:id:
    get:
      tags:
        - "Informatives"
      summary: "Get informative information"
      produces:
      - "application/json"
      parameters:
        - in: param
          name: id
          required: true
          type: string
          format: "uuid"
      responses:
        "200":
          description: "Obtained a informative"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              title:
                type: "string"
              content:
                type: "string"
        "500":
          description: "Could not obtain a informative"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    delete:
      tags:
        - "Informatives"
      summary: "Delete informative"
      produces:
      - "application/json"
      parameters:
        - in: param
          name: id
          required: true
          type: string
          format: "uuid"
      responses:
        "200":
          description: "Successfully removed a informative"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              title:
                type: "string"
              content:
                type: "string"
        "500":
          description: "Could not remove a informative"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
...